@@ -1,11 +1,55 @@
 /*!
- * Copyright (c) 2021-2022 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from 'bedrock';
+import https from 'https';
 
-// translate `main.js` to CommonJS
-require = require('esm')(module);
-module.exports = require('./main.js');
+// load config
+import './config.js';
 
-// backwards compatibility: export full API as `default`
-module.exports.default = module.exports;
+const {config, util: {BedrockError}} = bedrock;
+
+let _httpsAgent = null;
+
+bedrock.events.on('bedrock.init', () => {
+  // inialize httpsAgent after bedrock configuration stage is complete
+  _httpsAgent = new https.Agent(config['https-agent']);
+});
+
+export const httpsAgent = new Proxy({}, {
+  get(target, propKey) {
+    if(!_httpsAgent) {
+      throw _invalidStateError();
+    }
+    return _httpsAgent[propKey];
+  },
+  getPrototypeOf() {
+    if(!_httpsAgent) {
+      throw _invalidStateError();
+    }
+    return Object.getPrototypeOf(_httpsAgent);
+  },
+  getOwnPropertyDescriptor(target, prop) {
+    if(!_httpsAgent) {
+      throw _invalidStateError();
+    }
+    return Object.getOwnPropertyDescriptor(_httpsAgent, prop);
+  },
+  set(target, prop, value) {
+    if(!_httpsAgent) {
+      throw _invalidStateError();
+    }
+    _httpsAgent[prop] = value;
+    return true;
+  }
+});
+export const agent = httpsAgent;
+
+function _invalidStateError() {
+  return new BedrockError('The agent is not ready.', 'InvalidStateError');
+}
+
+// export for testing purposes only
+export function _setHttpsAgent({httpsAgent}) {
+  _httpsAgent = httpsAgent;
+}
